pyproject.toml
[tool.poetry]
name = "job-scheduler-experiment"
version = "0.0.1"
description = ""
authors = ["Petr Lavrov <calmmage@gmail.com>"]
readme = "README.md"
packages=[
    { include = "job_scheduler_experiment", from = "." },
]


[tool.poetry.dependencies]
python = ">=3.11,<4.0"
calmlib = { git = "https://github.com/calmmage/calmlib.git", branch = "main" }
loguru = ">=0.7"
# for parsing version
toml = ">=0.10"
python-dotenv = "*"
fastapi = "^0.115.12"
pymongo = "^4.12.0"
aiogram = "^3.20.0"
pydantic = "^2.11.3"
pydantic-settings = "^2.8.1"
typer = "^0.15.2"
rich = "^14.0.0"
uvicorn = "^0.34.1"
apscheduler = "^3.10.4"
motor = "^3.7.0"

[tool.poetry.group.extras.dependencies]
# dependencies for extra features

# Main
## Pydantic
#pydantic = ">=2"            # Data validation and settings management
#pydantic-settings = ">=2"   # Pydantic settings management

# git ops
#PyGithub = "*"              # GitHub API client
#gitpython = "*"             # Git repository interface

# api
#fastapi = "*"               # High-performance API framework
#httpx = "*"                 # Async HTTP client

# cli
#click = "*"                 # CLI creation toolkit
#typer = "*"                 # Easy CLI building

# data, databases
#pymongo = "*"                 # MongoDB client
#mongoengine = "*"            # MongoDB ORM
#PyDrive = "*"                # Google Drive API client
#protobuf = "*"               # Protocol buffers
#dropbox = "*"                # Dropbox API client

# utils
#pytz = "*"                  # Timezone definitions
#retry = "*"                 # Retry decorator
#nltk = "*"                  # Natural language toolkit
#easydict = ">=1.10"         # Easy access to dict keys - dotdict
#pydub = ">=0.25"             # Audio manipulation
#apscheduler = ">=3"         # Scheduling library
#pyperclip = "*"              # Copying to clipboard
#aiolimiter = ">=1.1"         # Async rate limiting
#deepl = ">=1.18"             # DeepL API client

# LLM, GPT, AI
#tiktoken = ">=0.7"            # Tokenization for OpenAI models
#openai = ">=1.45"             # OpenAI API client
#langchain = ">=0.3"            # Language model framework
#langchain-openai = ">=0.2"     # OpenAI integration for LangChain
#langchain-community = "*"      # Community extensions for LangChain
#langchain-anthropic = ">=0.2"  # Anthropic integration for LangChain
#langfuse = "*"                 # Language model analytics
#assemblyai = "*"              # Speech-to-text API
#whisperx = { git = "https://github.com/m-bain/whisperX.git", branch = "main" }  # Advanced Whisper model

# Misc
#ipykernel = "*"             # IPython kernel
#ipywidgets = "*"            # Interactive widgets for Jupyter
#joblib = "*"                # Lightweight pipelining
#paramiko = "*"              # SSH2 protocol library
#scp = "*"                   # Secure copy
#pymorphy2 = "*"             # Morphological analyzer

[tool.poetry.group.test.dependencies]
pytest = ">=6"
#pytest-mongodb = "*"
black = {extras = ["jupyter"], version = "^25.1.0"}
isort = "^6.0.1"
vulture = "^2.14"
pytest-cov = "^6.1.1"
pytest-asyncio = "^0.26.0"
pytest-mock = "^3.14.0"
pyupgrade = "^3.19.1"
pyright = "^1.1.399"
ruff = "^0.11.5"

[tool.poetry.group.docs.dependencies]
#mkdocs = ">=1.6"

[tool.poetry.group.dev.dependencies]
# dev / research utils
jupyter = ">=1.0.0"          # Interactive notebooks
#jupyter-nbextensions-configurator = "*"
#jupyter_contrib_nbextensions = "*"
black = ">=24.0.0"           # Code formatter
pre-commit = ">=3.8.0"       # Pre-commit hooks
tqdm = "*"                   # Progress bars

## data analysis, research
#pandas = ">=2.0.0"          # Data manipulation and analysis
#seaborn = "*"               # Statistical data visualization
#bokeh = "*"                 # Interactive visualizations
#matplotlib = "*"            # Plotting library
#numpy = "*"                 # Numerical computing

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[template]
version = "0.3.0"  # optimize pyproject dependencies
url = "https://github.com/calmmage/python-project-template.git"
author = "Petr Lavrov <petrlavrov@calmmage.com>"
description = "A template for creating a bot using calmapp and bot-lib"